Crate spl_memo

A program that accepts a string of encoded characters and verifies that it parses, while verifying and logging signers. Currently handles UTF-8 characters.

Re-exports

pub use solana_program;

Modules

processor

Program state processor

v1

Legacy symbols from Memo v1

Constants

ID

The const program ID.

Functions

build_memo

Build a memo instruction, possibly signed

check_id

Returns true if given pubkey is the program ID.

id

Returns the program ID.