Crate spl_memo

A program that accepts a string of encoded characters and verifies that it parses, while verifying and logging signers. Currently handles UTF-8 characters.

Re-exports

pub use solana_account_info;

pub use solana_instruction;

pub use solana_msg;

pub use solana_program_entrypoint;

pub use solana_program_error;

pub use solana_pubkey;

Modules

processor

Program state processor

v1

Legacy symbols from Memo v1

Constants

ID

The const program ID.

Functions

build_memo

Build a memo instruction, possibly signed

check_id

Returns true if given pubkey is the program ID.

id

Returns the program ID.